Remove build/run from Userdata, easy to update
AWS doesn't allow us to edit UserData, even on a copy of a Launch
Config. So do the minimum in the User Data and move build and run to
files we can easily update in the repo. Then all we have to do to change
the run/build procedures is update the code, push to GitHub, and kill
the server: AWS will relaunch and and the user data will find our
updated build and run scripts.